Refactor: Display user post count as diamonds in ProfilePage app bar
A new widget, `_buildDiamondCounter`, has been added to the `ProfilePage` to display the user's post count as a diamond icon followed by the count.

This diamond counter is now shown in the leading position of the `AppBar`.

The `leadingWidth` of the `AppBar` has been adjusted to accommodate this new widget.